requirements.txt
requests>=2.31.0
pandas>=2.2.0
PyYAML>=6.0.1
pulp>=2.8.0
matplotlib>=3.8.0
pytest>=8.0.0
